chore(api): split content.ts into per-resource route files
Break the 2.4k-line routes/content.ts into one file per /v1 resource
(posts, brand-identities, integrations, legacy-redirects) and extract
shared helpers into utils/. Pure refactor with no behavior change.
